# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Pessoas",
        "description": "Teste para a vaga de PHP Pleno",
        "version": "1.0.0"
    },
    "paths": {
        "/api/enderecos": {
            "get": {
                "tags": [
                    "Endereços"
                ],
                "summary": "Listar todos registros",
                "description": "",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Endereços"
                ],
                "summary": "Cadastrar endereço",
                "description": "",
                "operationId": "store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "logradouro": {
                                        "type": "string"
                                    },
                                    "complemento": {
                                        "type": "string"
                                    },
                                    "bairro": {
                                        "type": "string"
                                    },
                                    "uf": {
                                        "type": "string"
                                    },
                                    "cidade": {
                                        "required": [
                                            "uf"
                                        ],
                                        "type": "string"
                                    },
                                    "numero": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "cep": "95873-377",
                                    "logradouro": "Rua Sebastião Escobar, 17. Bc. 7 Ap. 50",
                                    "numero": "4557",
                                    "complemento": "Apto 89",
                                    "bairro": "Rua Perez",
                                    "pais": "Brasil",
                                    "uf": "AM",
                                    "cidade": "Carrara do Leste"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            }
        },
        "/api/enderecos/{id}": {
            "get": {
                "tags": [
                    "Endereços"
                ],
                "summary": "Exibir dados de um endereço",
                "description": "",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do endereço",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            },
            "put": {
                "tags": [
                    "Endereços"
                ],
                "summary": "Atualizar um endereço",
                "description": "",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do endereço",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "logradouro": {
                                        "type": "string"
                                    },
                                    "complemento": {
                                        "type": "string"
                                    },
                                    "bairro": {
                                        "type": "string"
                                    },
                                    "uf": {
                                        "type": "string"
                                    },
                                    "cidade": {
                                        "required": [
                                            "uf"
                                        ],
                                        "type": "string"
                                    },
                                    "numero": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "cep": "95873-377",
                                    "logradouro": "Rua Sebastião Escobar, 17. Bc. 7 Ap. 50",
                                    "numero": "4557",
                                    "complemento": "Apto 89",
                                    "bairro": "Rua Perez",
                                    "pais": "Brasil",
                                    "uf": "AM",
                                    "cidade": "Carrara do Leste"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Endereços"
                ],
                "summary": "Deletar um endereço",
                "description": "",
                "operationId": "delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do endereço",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
    }
}